feat: sbp function to join with invite secret
